fixes None type errors in autorooms

+++ autorooms/autorooms.py
@@ -99,11 +99,12 @@ class AutoRooms:
                     cache.append(channel.id)
                     self.save_json()
 
-        channel = memb_before.voice.voice_channel
-        if channel.id in cache:
-            if len(channel.voice_members) == 0:
-                await self.bot.delete_channel(channel)
-                self.settingscleanup(server)
+        if memb_before.voice.voice_channel is not None:
+            channel = memb_before.voice.voice_channel
+            if channel.id in cache:
+                if len(channel.voice_members) == 0:
+                    await self.bot.delete_channel(channel)
+                    self.settingscleanup(server)
 
     def settingscleanup(self, server):
         """cleanup of settings"""
